pyproject.toml
[project]
name = "airflow-dbt-cosmos-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "apache-airflow==2.10.0",
    "apache-airflow-providers-amazon>=9.0.0",
    "apache-airflow-providers-postgres>=5.11.3",
    "astronomer-cosmos>=1.12.1",
]
